테스트용

+++ src/main/java/com/threeteam/sns/dto/TracksDto.java
@@ -1,13 +1,15 @@
 package com.threeteam.sns.dto;
+
 import lombok.Data;
 
 @Data
 public class TracksDto {
-    private Long id;
+	private Integer id;
     private String preview;      // 미리듣기 URL
     private String titleShort;   // title_short
     private String albumCover;   // album.cover_medium
     private String albumTitle;   // album.title
     private String artistName;   // artist.name
     private int duration;        // 음악 길이 (초)
+    
 }
